Refactor classic battle helpers
